Adding S13 test

+++ impl/src/test/java/com/mobius/software/telco/protocols/diameter/test/load/LoadTest.java
@@ -225,7 +225,7 @@ public class LoadTest extends NetworkTestBase
 		//make sure no timeout is processed
 		try
 		{
-			Thread.sleep(idleTimeout * 2);
+			Thread.sleep((int)(idleTimeout * 2));
 		}
 		catch(InterruptedException ex)
 		{
